Refactor location of code
1. The 'cayley' command now lives under cmd/
2. Integration tests live under integration/
3. Some utility functions used in both the integration tests and the
   'cayley' command were split to their own 'util' package.
