+++ cmd/cayley/cayley.go
@@ -19,11 +19,7 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io"
-	client "net/http"
-	"net/url"
 	"os"
-	"path/filepath"
 	"runtime"
 	"time"
 
@@ -33,10 +29,7 @@ import (
 	"github.com/google/cayley/db"
 	"github.com/google/cayley/graph"
 	"github.com/google/cayley/http"
-	"github.com/google/cayley/internal"
-	"github.com/google/cayley/quad"
-	"github.com/google/cayley/quad/cquads"
-	"github.com/google/cayley/quad/nquads"
+	"github.com/google/cayley/util"
 
 	// Load all supported backends.
 	_ "github.com/google/cayley/graph/bolt"
@@ -192,7 +185,7 @@ func main() {
 			if err != nil {
 				break
 			}
-			err = load(handle.QuadWriter, cfg, *quadFile, *quadType)
+			err = util.Load(handle.QuadWriter, cfg, *quadFile, *quadType)
 			if err != nil {
 				break
 			}
@@ -204,7 +197,7 @@ func main() {
 		if err != nil {
 			break
 		}
-		err = load(handle.QuadWriter, cfg, *quadFile, *quadType)
+		err = util.Load(handle.QuadWriter, cfg, *quadFile, *quadType)
 		if err != nil {
 			break
 		}
@@ -217,7 +210,7 @@ func main() {
 			break
 		}
 		if !graph.IsPersistent(cfg.DatabaseType) {
-			err = load(handle.QuadWriter, cfg, "", *quadType)
+			err = util.Load(handle.QuadWriter, cfg, "", *quadType)
 			if err != nil {
 				break
 			}
@@ -233,7 +226,7 @@ func main() {
 			break
 		}
 		if !graph.IsPersistent(cfg.DatabaseType) {
-			err = load(handle.QuadWriter, cfg, "", *quadType)
+			err = util.Load(handle.QuadWriter, cfg, "", *quadType)
 			if err != nil {
 				break
 			}
@@ -250,64 +243,4 @@ func main() {
 	if err != nil {
 		glog.Errorln(err)
 	}
-}
-
-func load(qw graph.QuadWriter, cfg *config.Config, path, typ string) error {
-	return decompressAndLoad(qw, cfg, path, typ, db.Load)
-}
-
-func decompressAndLoad(qw graph.QuadWriter, cfg *config.Config, path, typ string, loadFn func(graph.QuadWriter, *config.Config, quad.Unmarshaler) error) error {
-	var r io.Reader
-
-	if path == "" {
-		path = cfg.DatabasePath
-	}
-	if path == "" {
-		return nil
-	}
-	u, err := url.Parse(path)
-	if err != nil || u.Scheme == "file" || u.Scheme == "" {
-		// Don't alter relative URL path or non-URL path parameter.
-		if u.Scheme != "" && err == nil {
-			// Recovery heuristic for mistyping "file://path/to/file".
-			path = filepath.Join(u.Host, u.Path)
-		}
-		f, err := os.Open(path)
-		if err != nil {
-			return fmt.Errorf("could not open file %q: %v", path, err)
-		}
-		defer f.Close()
-		r = f
-	} else {
-		res, err := client.Get(path)
-		if err != nil {
-			return fmt.Errorf("could not get resource <%s>: %v", u, err)
-		}
-		defer res.Body.Close()
-		r = res.Body
-	}
-
-	r, err = internal.Decompressor(r)
-	if err != nil {
-		if err == io.EOF {
-			return nil
-		}
-		return err
-	}
-
-	var dec quad.Unmarshaler
-	switch typ {
-	case "cquad":
-		dec = cquads.NewDecoder(r)
-	case "nquad":
-		dec = nquads.NewDecoder(r)
-	default:
-		return fmt.Errorf("unknown quad format %q", typ)
-	}
-
-	if loadFn != nil {
-		return loadFn(qw, cfg, dec)
-	}
-
-	return db.Load(qw, cfg, dec)
 }
